Main activity: fix overlapping snackbar and FAB
If you view the diff with whitespace ignored, then the fix is rather
simple: just need to wrap the whole layout into a coordinator layout,
then move the FloatingActionButton out to the CoordinatorLayout and that
fixes the problem out of the box.

See
<https://stackoverflow.com/questions/40929686/constraintlayout-vs-coordinator-layout>
for details, coordinator and constraint layout is not an either/or, they
can be combined.

Fixes <https://github.com/vmiklos/plees-tracker/issues/8>.

Tested-by: Travis