Fix white background of the ActionBar
Which is problematic, since the text is white as well.

Just don't customize this: this way the ActionBar background is purple
in light mode, black in dark mode. Also, this way the dark mode
background is not exactly black, just some gray, similar to what other
apps do.

Related to <https://github.com/vmiklos/plees-tracker/issues/16>.

Tested-by: Travis