Add a setting to create automatic exports for backups
The setting is off by default.

Once it's on, the user can choose a folder. The backup.csv in that
folder is updated each time a sleep is stopped. Keep it simple like like
for now, we could later limit this to "once a week", if it's needed.

Fixes <https://github.com/vmiklos/plees-tracker/issues/32>.

Change-Id: I341b3ce1dbdf29b4c4a73b8f4015fda4b1d132e5