Simplify readme, add user doc
It was odd to specify the import/export format in the feature summary.
And lots of other details were only visible to the user in case their
experimented or read the code.

Not to mention that the test coverage is quite poor: so at least
documenting the inteded behavior is a good step forward.

Change-Id: I586104a2375be0949dc21c8d8cc639ce5506383a