Fix issues with calendar import creating duplicates or not importing
Instead of checking against the last time an import was made, we just
ignore any duplicate sleeps imported. This is a more robust solution
since importing works as expected even if app data is deleted or we
change devices.

This method also allows for additions and changes to calendar sleep
events in the past from the last import to get picked up. For
example if you make manual changes, or use the app on more than one
device but don't regularly export to calendar.