Fix lack of toggle from the notification drawer when app is already foreground
The steps from
<https://github.com/vmiklos/plees-tracker/issues/292#issuecomment-1185306926>
describe that in case the app is already in the foreground, then stop
from the notification drawer won't do anything.

The trouble is that the toggle is implemented as a side-effect of
the creation of the main activity, which is optimized away when the main
activity is already the top of the activity stack. The benefit of this
approach is that once the toggle happens from the main activity, we can
be sure that the DataModel singleton is initialized and also can use the
view model of the activity for coroutine purposes (the stop involves an
async db operation).

Fix the problem by following the advice from
<https://stackoverflow.com/questions/53459111/launch-activity-from-notification-instantiating-activity-even-if-already-there>:

- set the activity's launch mode to singleTop

- implement the new onNewIntent(), which will be called in the case
  where nothing happened previously

Note that toggle from the widget didn't have this problem, since you can
only see the widget when the activity is not at the top of the activity
stack.

Fixes <https://github.com/vmiklos/plees-tracker/issues/292>.

Change-Id: I6861673c0ada5c74d80654bebbc78f98fbb3aa8f