Don't require app restart when changing the `use_median` setting
Have the getAfterLive() and the liveDataBoolean("use_median", ...) calls
in separate scopes, so in case we want to add more settings there in the
future, then no need to increase nesting again this way.

And in case afterSleep is not a callable, then it would be only called
once, so the chained LiveData wouldn't update properly.

Fixes <https://github.com/vmiklos/plees-tracker/issues/474>.