Export to file: add a pretty CSV mode
End-users sometimes want to show their sleep log to other people, but
the data exported to a file, but that is not human-readable.

The default export format is a CSV that is meant to be easy to import
back, i.e. a machine-readable file format.

Fix the problem by adding a setting (defaults to false): if enabled, the
export result will have formatted timestamps and also the sleep length
will be written there, if if that's redundant for a machine reader.

Fixes <https://github.com/vmiklos/plees-tracker/issues/515>.