Enable edge-to-edge mode and fix MainActivity
Android 15 defaulted to this, but we could opt out. Android 16 will
ignore the opt out, see
<https://developer.android.com/develop/ui/views/layout/edge-to-edge>.

Once edge-to-edge mode is enabled, two things break: the root view
overlaps with the top status bar and the (optional) bottom navigator
bar; also once the necessary insets are applied so this doesn't happen,
the top status bar will have light text on light background, which is
not readable.

Fix the first by using setOnApplyWindowInsetsListener() on the root view
of the main activity; fix the second by enabling
isAppearanceLightStatusBars when we're in light mode.

The other 4 activities still need similar fixing. One nice detail is
that the activity is restarted when dark mode is toggled, so no app
restart is needed to have correct light/dark text in the top status bar.